requirements.txt
apache-flink==1.17.1
kafka-python==2.0.2
clickhouse-driver==0.2.6